move tier hologram to v2
